refactor(ui): split cars wizard view
