+++ apps/ui/README.md
@@ -123,8 +123,10 @@ source-of-truth export commands remain the only writers for those files.
 | `app/views/update_panel.tsx` | Signal-backed Preact owner for the full update settings surface that renders the action row plus current status, health, journey, issues, latest attempt, and log cards through a semantic panel bridge |
 | `app/views/sensors_panel.tsx` | Signal-backed Preact owner for the sensors settings shell that keeps the sensor table reactive while exposing typed identify/remove/location callbacks to the realtime feature |
 | `app/views/speed_source_panel.tsx` | Signal-backed Preact owner for the speed-source shell that renders the full tab plus live diagnostics in JSX, owns typed save/scan/select/input callbacks, and exposes semantic `setModel()` / `setDiagnostics()` bridge updates to the speed-source and GPS-status modules |
-| `app/views/cars_panel.tsx` | Signal-backed Preact owner for the full car-management surface; it renders saved-car guidance/list rows plus the full add-car wizard in JSX, owns wizard focus/return-focus/scroll lifecycle locally, and exposes typed list and wizard bridges |
-| `app/views/cars_feature_presenter.ts` | Thin car-wizard presenter that turns workflow state into typed wizard render models and delegates focus/manual-input access through the island bridge |
+| `app/views/cars_panel.tsx` | Signal-backed Preact owner for the full car-management surface; it renders saved-car guidance/list rows, delegates wizard focus/return-focus/scroll lifecycle to the extracted wizard focus hook, and exposes typed list and wizard bridges |
+| `app/views/cars_wizard_focus.ts` | Focus/ref owner for the add-car wizard that centralizes return-focus, scroll reset, requested-focus handling, and target resolution now that the wizard lifecycle lives beside the cars panel surface |
+| `app/views/cars_wizard_panel.tsx` | Modal shell for the add-car wizard; it keeps dialog chrome and typed action wiring small while delegating step content to focused wizard sections |
+| `app/views/cars_wizard_sections.tsx` | Extracted add-car wizard step sections, option grids, manual-spec inputs, and summary helpers that keep the main wizard panel readable while preserving the existing selectors and flow |
 | `app/views/car_wizard_view.ts` | Typed add-car wizard render-model builders for progress, option sections, selected specs, and summary rows reused by the Preact car-management island |
 | `app/features/update_feature.ts` | Thin update facade that binds typed island actions, delegates island render-model updates to the presenter, and derives update/internet polling context from the shell and settings tab state |
 | `app/features/update_feature_workflow.ts` | DOM-free update workflow/controller for update polling, internet-status normalization, and start/cancel command orchestration |
