+++ apps/ui/src/app/views/cars_panel.tsx
@@ -1,10 +1,8 @@
 import { render } from "preact";
-import { useRef } from "preact/hooks";
 
 import type { CarsFeatureFocusTarget } from "../features/cars_feature_workflow";
 import {
   signal,
-  useSignalEffect,
   type ReadonlySignal,
 } from "../ui_signals";
 import {
@@ -18,13 +16,13 @@ import {
 } from "./cars_list_section";
 import {
   CarsWizardPanel,
-  resolveWizardFocusTarget,
   type CarsFeatureInteraction,
   type CarsFeatureInteractionHandlers,
-  type CarsWizardElementRefs,
-  type CarsWizardFocusRequest,
-  type CarsWizardOptionRefs,
 } from "./cars_wizard_panel";
+import {
+  useCarsWizardFocusManager,
+  type CarsWizardFocusRequest,
+} from "./cars_wizard_focus";
 
 export type { CarsFeatureInteraction, CarsFeatureInteractionHandlers } from "./cars_wizard_panel";
 export type { CarsListRenderModel } from "./cars_list_section";
@@ -57,10 +55,6 @@ const DEFAULT_CARS_PANEL_MODEL: CarsListRenderModel = {
   table: null,
 };
 
-function focusElement(target: HTMLElement | null | undefined): void {
-  target?.focus();
-}
-
 function CarsPanel(props: {
   state: ReadonlySignal<CarsPanelBridgeState>;
   wizardFocusRequest: ReadonlySignal<CarsWizardFocusRequest | null>;
@@ -68,104 +62,16 @@ function CarsPanel(props: {
   const state = props.state.value;
   const model = state.model?.value ?? DEFAULT_CARS_PANEL_MODEL;
   const wizardModel = state.wizardModel?.value ?? createClosedCarsWizardRenderModel();
-  const addCarBtnRef = useRef<HTMLButtonElement | null>(null);
-  const addCarWizardRef = useRef<HTMLDivElement | null>(null);
-  const wizardCloseBtnRef = useRef<HTMLButtonElement | null>(null);
-  const wizardCustomBrandInputRef = useRef<HTMLInputElement | null>(null);
-  const wizardCustomModelInputRef = useRef<HTMLInputElement | null>(null);
-  const wizardCustomTypeInputRef = useRef<HTMLInputElement | null>(null);
-  const wizardManualAddBtnRef = useRef<HTMLButtonElement | null>(null);
-  const wizFinalDriveInputRef = useRef<HTMLInputElement | null>(null);
-  const wizGearRatioInputRef = useRef<HTMLInputElement | null>(null);
-  const wizRimInputRef = useRef<HTMLInputElement | null>(null);
-  const wizTireAspectInputRef = useRef<HTMLInputElement | null>(null);
-  const wizTireWidthInputRef = useRef<HTMLInputElement | null>(null);
-  const optionRefs = useRef<CarsWizardOptionRefs>({
-    brandOption: null,
-    gearboxOption: null,
-    modelOption: null,
-    tireOption: null,
-    typeOption: null,
-    variantOption: null,
+  const { addCarButtonRef, wizardRefs } = useCarsWizardFocusManager({
+    state: props.state,
+    wizardFocusRequest: props.wizardFocusRequest,
   });
-  const lastReturnFocusTargetRef = useRef<HTMLElement | null>(null);
-  const lastHandledFocusRequestTokenRef = useRef(0);
-  const lastWizardOpenStateRef = useRef(wizardModel.isOpen);
-
-  useSignalEffect(() => {
-    const isOpen = props.state.value.wizardModel?.value.isOpen ?? false;
-    const wasOpen = lastWizardOpenStateRef.current;
-    if (isOpen && !wasOpen) {
-      const activeElement = document.activeElement instanceof HTMLElement
-        ? document.activeElement
-        : null;
-      lastReturnFocusTargetRef.current =
-        activeElement && activeElement !== document.body ? activeElement : addCarBtnRef.current;
-      queueMicrotask(() => {
-        if (addCarWizardRef.current) {
-          addCarWizardRef.current.scrollTop = 0;
-        }
-      });
-    }
-    if (!isOpen && wasOpen) {
-      const target = lastReturnFocusTargetRef.current;
-      lastReturnFocusTargetRef.current = null;
-      queueMicrotask(() => {
-        const safeTarget = target && document.contains(target) ? target : addCarBtnRef.current;
-        focusElement(safeTarget);
-      });
-    }
-    lastWizardOpenStateRef.current = isOpen;
-  });
-
-  useSignalEffect(() => {
-    const wizardFocusRequest = props.wizardFocusRequest.value;
-    if (
-      !wizardFocusRequest ||
-      wizardFocusRequest.token === lastHandledFocusRequestTokenRef.current
-    ) {
-      return;
-    }
-    lastHandledFocusRequestTokenRef.current = wizardFocusRequest.token;
-    queueMicrotask(() => {
-      focusElement(
-        resolveWizardFocusTarget(wizardFocusRequest.target, {
-          closeButton: wizardCloseBtnRef.current,
-          customBrandInput: wizardCustomBrandInputRef.current,
-          customModelInput: wizardCustomModelInputRef.current,
-          customTypeInput: wizardCustomTypeInputRef.current,
-          finalDriveInput: wizFinalDriveInputRef.current,
-          manualAddButton: wizardManualAddBtnRef.current,
-          optionRefs: optionRefs.current,
-          rimInput: wizRimInputRef.current,
-          tireAspectInput: wizTireAspectInputRef.current,
-          tireWidthInput: wizTireWidthInputRef.current,
-          topGearInput: wizGearRatioInputRef.current,
-        }),
-      );
-    });
-  });
-
-  const wizardRefs: CarsWizardElementRefs = {
-    addCarWizardRef,
-    optionRefs,
-    wizardCloseBtnRef,
-    wizardCustomBrandInputRef,
-    wizardCustomModelInputRef,
-    wizardCustomTypeInputRef,
-    wizardManualAddBtnRef,
-    wizFinalDriveInputRef,
-    wizGearRatioInputRef,
-    wizRimInputRef,
-    wizTireAspectInputRef,
-    wizTireWidthInputRef,
-  };
 
   return (
     <>
       <CarsListSection
         actions={state.actions}
-        addCarButtonRef={addCarBtnRef}
+        addCarButtonRef={addCarButtonRef}
         model={model}
         onOpenWizard={() => state.wizardActions?.onAction({ type: "open" })}
       />
